Refactoring and requested changes
