+++ packages/webapp/src/components/LocationDetailLayout/PointDetails/Sensor/index.jsx
