+++ packages/webapp/src/components/LocationDetailLayout/PointDetails/Sensor/styles.module.scss
