+++ packages/webapp/src/containers/LocationDetails/PointDetails/SensorDetail/index.jsx
@@ -12,12 +12,11 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
-import PureSensorDetail from '../../../../components/LocationDetailLayout/Sensor/SensorDetail/index';
+import PureSensorDetail from '../../../../components/LocationDetailLayout/PointDetails/Sensor/index';
 import { measurementSelector } from '../../../userFarmSlice';
-import { tasksFilterSelector } from '../../../filterSlice';
 import { useSelector, useDispatch } from 'react-redux';
 import { useTranslation } from 'react-i18next';
-import { useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 import { sensorsSelector } from '../../../sensorSlice';
 import { isAdminSelector } from '../../../userFarmSlice';
 import { getSensorReadingTypes, getSensorBrand } from './saga';
@@ -26,18 +25,14 @@ export default function SensorDetail({ history, user, match }) {
   const dispatch = useDispatch();
   const location_id = match.params.location_id;
   const sensorInfo = useSelector(sensorsSelector(location_id));
-
+  const system = useSelector(measurementSelector);
+  const isAdmin = useSelector(isAdminSelector);
   const { sensor_id, partner_id } = sensorInfo;
+
   useEffect(() => {
     dispatch(getSensorReadingTypes({ location_id, sensor_id }));
     dispatch(getSensorBrand({ location_id, partner_id }));
   });
-
-  const system = useSelector(measurementSelector);
-
-  const { t } = useTranslation();
-
-  const isAdmin = useSelector(isAdminSelector);
 
   return (
     <PureSensorDetail history={history} isAdmin={isAdmin} system={system} sensorInfo={sensorInfo} />
