+++ packages/webapp/src/routes/SensorDetailsRoutes.jsx
@@ -1,16 +1,18 @@
 import { Route } from 'react-router-dom';
 import React from 'react';
-import LocationTasks from '../containers/LocationDetails/LocationTasks';
 import SensorDetail from '../containers/LocationDetails/PointDetails/SensorDetail';
 import UpdateSensor from '../containers/LocationDetails/PointDetails/SensorDetail/EditSensor';
+import { isAdminSelector } from '../containers/userFarmSlice';
+import { useSelector } from 'react-redux';
 
 export default function SensorDetailsRoutes() {
+  const isAdmin = useSelector(isAdminSelector);
   return (
     <>
       {/* <Route path="/sensor/:location_id/tasks" exact component={LocationTasks} /> */}
       {/* <Route path="/sensor/:location_id/readings" exact component={SensorReadings} /> */}
       <Route path="/sensor/:location_id/details" exact component={SensorDetail} />
-      <Route path="/sensor/:location_id/edit" exact component={UpdateSensor} />
+      {isAdmin && <Route path="/sensor/:location_id/edit" exact component={UpdateSensor} />}
     </>
   );
 }
